requirements.txt
Flask==3.0.0
numpy==2.1.0
Pillow==10.4.0
gunicorn==21.2.0
